# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-SERVICES marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}/accept:
    post:
      operationId: acceptDelivery
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/request-revision:
    post:
      operationId: requestRevision
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /payments/checkout:
    post:
      operationId: createCheckout
      tags: [payments]
      summary: Create a DiamoPay payment link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
  /payments/{orderId}:
    get:
      operationId: getPayment
      tags: [payments]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # REVIEWS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isBanned: { type: boolean }
        createdAt: { type: string }
...
    FreelancerProfile:
      type: object
      required: [id, userId, level, totalEarnings, completedOrders, rating, reviewCount, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        level: { type: string }
        skills: { type: string }
        totalEarnings: { type: number }
        completedOrders: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        isAvailable: { type: boolean }
        createdAt: { type: string }
        user:
          $ref: "#/components/schemas/User"
...
    Category:
      type: object
      required: [id, name, slug, icon, serviceCount]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        icon: { type: string }
        description: { type: ["string", "null"] }
        serviceCount: { type: integer }
...
    Service:
      type: object
      required: [id, title, description, price, deliveryDays, categoryId, freelancerId, status, orderCount, rating, reviewCount, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        deliveryDays: { type: integer }
        categoryId: { type: integer }
        freelancerId: { type: integer }
        imageUrl: { type: ["string", "null"] }
        tags: { type: ["string", "null"] }
        status: { type: string }
        orderCount: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        isFeatured: { type: boolean }
        createdAt: { type: string }
        category:
          $ref: "#/components/schemas/Category"
        freelancer:
          $ref: "#/components/schemas/FreelancerProfile"
...
    Order:
      type: object
      required: [id, serviceId, clientId, freelancerId, status, amount, commission, freelancerAmount, description, deliveryDays, deadline, createdAt]
      properties:
        id: { type: integer }
        serviceId: { type: integer }
        clientId: { type: integer }
        freelancerId: { type: integer }
        status: { type: string }
        amount: { type: number }
        commission: { type: number }
        freelancerAmount: { type: number }
        description: { type: string }
        attachmentUrl: { type: ["string", "null"] }
        deliveryUrl: { type: ["string", "null"] }
        deliveryDays: { type: integer }
        deadline: { type: string }
        revisionNote: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        service:
          $ref: "#/components/schemas/Service"
        client:
          $ref: "#/components/schemas/User"
        freelancer:
          $ref: "#/components/schemas/FreelancerProfile"
    OrderInput:
      type: object
      required: [serviceId, description]
      properties:
        serviceId: { type: integer }
        description: { type: string }
        attachmentUrl: { type: string }
...
    RevisionInput:
      type: object
      required: [note]
      properties:
        note: { type: string }
...
    CheckoutInput:
      type: object
      required: [orderId]
      properties:
        orderId: { type: integer }
    CheckoutResponse:
      type: object
      required: [paymentUrl, reference]
      properties:
        paymentUrl: { type: string }
        reference: { type: string }
    Payment:
      type: object
      required: [id, orderId, amount, status, reference, createdAt]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        amount: { type: number }
        status: { type: string }
        reference: { type: string }
        provider: { type: string }
        createdAt: { type: string }
...