# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-SERVICES marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { FreelancerProfile } from './freelancerProfile';
import type { Service } from './service';
import type { User } from './user';

export interface Order {
  id: number;
  serviceId: number;
  clientId: number;
  freelancerId: number;
  status: string;
  amount: number;
  commission: number;
  freelancerAmount: number;
  description: string;
  /** @nullable */
  attachmentUrl?: string | null;
  /** @nullable */
  deliveryUrl?: string | null;
  deliveryDays: number;
  deadline: string;
  /** @nullable */
  revisionNote?: string | null;
  createdAt: string;
  updatedAt?: string;
  service?: Service;
  client?: User;
  freelancer?: FreelancerProfile;
}